Rename base_decorator.rb to decorators.rb. Add CapitalizeDecorator and TrimmerDecorator.
