--- base_decorator.rb
@@ -1,12 +1,0 @@
-require_relative 'class_nameable'
-
-class BaseDecorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
